fix: fix error in console
